move comment metadata to its own file
