+++ types/schemas/postSubmissionApplication/data/Comment.ts
@@ -1,4 +1,3 @@
-import {DateTime} from '../../../shared/utils';
 import {PublicCommentSummary, SpecialistCommentSummary} from './CommentSummary';
 import {PublicComment, PublicCommentRedacted} from './PublicComment';
 import {Specialist, SpecialistRedacted} from './SpecialistComment';
@@ -62,22 +61,3 @@ export type SpecialistCommentsRedacted = {
    */
   comments: SpecialistRedacted[];
 };
-
-/**
- * @id #CommentMetaData
- * @description Metadata about a comment
- */
-export interface CommentMetaData {
-  /**
-   * When comment was submitted
-   */
-  submittedAt: DateTime;
-  /**
-   * When comment was validated
-   */
-  validatedAt?: DateTime;
-  /**
-   * When comment was made public
-   */
-  publishedAt?: DateTime;
-}
